Finished All Coins Components Layout
Discarded the modal, because it was overly compicated and cumbersome, with refs which wasn't working great with func components.
Built it with Material UI's ClickAwayListener, which worked pretty great.

Also made a GeneralTab and GeneralWindow components for this and further usage.
Layout for Energy SHOULD be very easy from this point.