Cleaned up after last commit's mess + uninstalled Material-UI (at last)
last commit had a bit of a mess going.
In this commit:
 - Rewrote the whole 'customizedWindowStyle' in GeneralWindow component as a composedClassName, and moved the css to the according css file as it should be.
 - Moved the contentContainer-div's style to generalWindow.css instead of it being passed to its children repeatedly.
 - Got rid of the last but of Material-UI that was still and CoinsTab and EnergyTab.
 - Uninstalled Material-UI Module (Finally!!)

side note:
this took about an hour, which I was panning to dedicate to creating the dnd part of ActionsTab, and instead I was forced by my past self to clean up bad work.  it ain't fun.

Don't compromise on doing a good and clean work before calling it done.
and if you don't have the time in one session -
First of all -you SHOULD plan better than that next time, so you wont leave mess behind.
and once it happened - commit anyway, and state the exact mess the code is currently in, preferably with a to-do list.